LRSD-7791 Rename components
