+++ workspaces/liferay-customer-workspace/client-extensions/liferay-customer-custom-element/src/components/BadgeButton/index.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export {default} from './DropDownWithDrillDown';
+export {default} from './BadgeButton';
