+++ workspaces/liferay-customer-workspace/client-extensions/liferay-customer-custom-element/src/components/Filter/components/FilterCheckbox/utils/getAvailableFieldsCheckboxs.jsx
