+++ workspaces/liferay-customer-workspace/client-extensions/liferay-customer-custom-element/src/components/Table/TableHeader/TableHeader.css
